LayoutTests/http/tests/cache/disk-cache/disk-cache-request-max-stale.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var testMatrix =
[
 [
  { responseHeaders: {'Cache-control': 'max-age=0', 'ETag': 'match' } },
  { responseHeaders: {'Cache-control': 'max-age=100' } },
  ],
 [
   {},
   { responseHeaders: {'Age': '200' } },
  ],
 [
  { requestHeaders: {'Cache-control': 'max-stale=0' } },
  { requestHeaders: {'Cache-control': 'max-stale' } },
  { requestHeaders: {'Cache-control': 'max-stale=100' } },
  ],
 [
  {},
  { requestHeaders: {'Cache-control': 'max-age=0' } },
  { requestHeaders: {'Cache-control': 'max-age=100' } },
  ],
 ];

description("Test for 'max-stale' request header");

var tests = generateTests(testMatrix);

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
